as: fixes
